feat: 오디오 부분 분리
