+++ src/components/audio-upload/styles.module.scss
@@ -1,0 +1,24 @@
+@import 'src/theme/variables.scss';
+@import 'src/theme/utils.scss';
+
+
+.image{
+  padding-left: px(15);
+  width:100%;
+  text-align: center;
+  @include flexLayout(row, center, flex-start, 6);
+}
+
+.removeImage{
+  margin-top: px(15);
+}
+
+.uploadedImage{
+  width: auto;
+  margin-top: px(15);
+  max-width:px(350);
+  max-height: px(160);
+}
+.disable {
+  display: none;
+}
