+++ src/hooks/useFormData.ts
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { FieldErrors, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as Yup from 'yup';
-import React from 'react';
+import React, { useRef } from 'react';
 import { queryClient } from '../apis/queryClient';
 import { QUERY_KEY } from '../apis/queryKeys';
 import { PATH } from '../routes/path';
@@ -17,6 +17,7 @@ interface IFormValues {
   fileInputRef?: React.RefObject<HTMLInputElement>;
   recodeRef?: React.RefObject<MediaRecorder | null>;
   audioChunks?: Blob[];
+  audioButtonRef?: React.RefObject<HTMLButtonElement>;
 }
 
 const defaultValues = {
@@ -26,6 +27,7 @@ const defaultValues = {
   fileInputRef: { current: null },
   recodeRef: { current: null },
   audioChunks: [],
+  audioButtonRef: { current: null },
 };
 
 interface props {
